Support for site admin "Profile Mode".
Profile Mode happens when a user is a member of a site, but cannot see other users within that site. In this instance, WordPress replaces the "Users" top level menu with "Profile", and traditionally uses 'profile.php' to handle this.

For our needs, this adds several bits of complexity:
* add_submenu_page() does not pass a user ID into current_user_can() checks, resulting in botched look-ups
* The root level profile.php cannot be bypassed without passing a query variable, and superfluous query variables will muck up metabox registration, caps checks, and other essential integrations
* Setting caps to `read` or `exist` are not very reliable, without intelligent mappings.

So, we refactor things a bit to work seamlessly with all user-edit scenarios:
* User can list_users, editing themself or someone else
* User cannot list_users, editing only themself
* In multisite, user can access network admin, editing themself or someone else
* In multisite, user accesses the User Dashboard, editing only themself