Permissions: normalize for_site() usage for _metabox() and save()
This change fixes a bug causing User Roles to be incorrectly set when saving the "Permissions" tab on multisite installations.

It also shifts the order of switch_to_blog() and current_user_can() calls, to ensure that Roles are only saved when the current user can "promote_users" on the current site for the user being edited (never self!)

It also also includes a little bit of surrounding code clean-up while I'm here.

Props chrisvanpatten. Fixes #70.